move socket.error classes from _socket to socket
They're implemented in C, but consider themselves to live in
socket.
